+++ extensions/polymarket/CHANGELOG.md
@@ -1,9 +1,9 @@
 # Polymarket Changelog
 
+## [PolyMarket Search] - {PR_MERGE_DATE}
+- Added ability to search PolyMarket for a specific market.
+
 ## [Remove old election markets] - 2024-11-18
 Removed 2024 US Election Odds and Swing States
 
 ## [Initial Version] - 2024-11-05
-
-## [PolyMarket Search] 2025 -05-10
-- Added ability to search PolyMarket for a specific market.